Testing Document
Here's a rough draft of the Testing Document. Some of the information contained in the document is the potential methods we could use to help us develop tests for our code. This is meant to be a rough draft and will be revised as we continue working on our project.